chore: better shared layer (with models, extensions and enums separateds)
