+++ lib/shared/enums/level_enum.dart
@@ -1,0 +1,6 @@
+enum Level {
+  easy,
+  middle,
+  hard,
+  expert,
+}
