+++ lib/shared/extensions/level_string_extension.dart
@@ -1,0 +1,12 @@
+import 'package:devquiz/shared/enums/level_enum.dart';
+
+extension LevelStringExtension on String {
+  Level get levelMapping =>
+      {
+        'easy': Level.easy,
+        'middle': Level.middle,
+        'hard': Level.hard,
+        'expert': Level.expert
+      }[this] ??
+      Level.easy;
+}
